requirements.txt
django~=5.2.4
requests
SQLAlchemy~=2.0.41
h11~=0.16.0
pip~=25.1.1
distro~=1.9.0
sniffio~=1.3.1
httpcore~=1.0.9
idna~=3.10
httpx~=0.28.1
asyncpg~=0.30.0
typing_extensions~=4.14.1
certifi~=2025.7.14
greenlet~=3.2.3
anyio~=4.9.0
psycopg2~=2.9.10
psycopg2-binary~=2.9.10
python-telegram-bot~=22.3